+++ src/ComputeFleet/ComputeFleet.Autorest/test/Remove-AzComputeFleet.Tests.ps1
@@ -15,20 +15,93 @@ if(($null -eq $TestName) -or ($TestName -contains 'Remove-AzComputeFleet'))
 }
 
 Describe 'Remove-AzComputeFleet' {
+
+    BeforeAll {
+        $resourceGroupName = "fleet-remove-test-rg"
+        $vnetName = "vnet"
+        $nsgName = "nsg"
+        $vmNamePrefix = "fleetvm"
+        $deleteFleetName = "delete-fleet"
+        $deleteViaIdentityFleetName = "delete-via-identity-fleet"
+
+        if ($TestMode -ne 'playback') {
+            $result = New-TestResourceGroup -ResourceGroupName $resourceGroupName `
+                -Location $env.Location -VNetName $vnetName -NsgName $nsgName
+            $subnetId = $result.SubnetId
+            $nsgId = $result.NsgId
+        } else {
+            $subnetId = "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test/providers/Microsoft.Network/virtualNetworks/vnet/subnets/subnet1"
+            $nsgId = "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test/providers/Microsoft.Network/networkSecurityGroups/nsg"
+        }
+    }
+
     It 'Delete' {
         {
-            Remove-AzComputeFleet -SubscriptionId $env.SubscriptionId -ResourceGroupName $env.ResourceGroupName -FleetName $env.FleetName2
-            $fleetList = Get-AzComputeFleet -SubscriptionId $env.SubscriptionId
-            $fleetList.Name | Should -Not -Contain $env.FleetName2
+            $vmProfile = New-TestVmProfile -SubnetId $subnetId -NsgId $nsgId
+            $vmSize = [Microsoft.Azure.PowerShell.Cmdlets.ComputeFleet.Models.VMSizeProfile]::new()
+            $vmSize.Name = "Standard_D2s_v3"
+
+            $fleet = New-AzComputeFleet -Name $deleteFleetName `
+                -ResourceGroupName $resourceGroupName `
+                -SubscriptionId $env.SubscriptionId `
+                -Location $env.Location `
+                -ComputeProfileBaseVirtualMachineProfile $vmProfile `
+                -ComputeProfileComputeApiVersion "2024-03-01" `
+                -VMSizesProfile @($vmSize) `
+                -RegularPriorityProfileCapacity 1 `
+                -RegularPriorityProfileMinCapacity 1 `
+                -RegularPriorityProfileAllocationStrategy "LowestPrice" `
+                -Mode "Launch" `
+                -VMNamePrefix $vmNamePrefix `
+                -Tag @{ environment = "test" }
+
+            $fleet.Name | Should -Be $deleteFleetName
+            $fleet.ProvisioningState | Should -Be "Succeeded"
+
+            Remove-AzComputeFleet -Name $deleteFleetName `
+                -ResourceGroupName $resourceGroupName `
+                -SubscriptionId $env.SubscriptionId
+
+            { Get-AzComputeFleet -Name $deleteFleetName `
+                -ResourceGroupName $resourceGroupName `
+                -SubscriptionId $env.SubscriptionId } | Should -Throw
         } | Should -Not -Throw
     }
 
     It 'DeleteViaIdentity' {
         {
-            $fleet = Get-AzComputeFleet -SubscriptionId $env.SubscriptionId -ResourceGroupName $env.ResourceGroupName -FleetName $env.FleetName3
+            $vmProfile = New-TestVmProfile -SubnetId $subnetId -NsgId $nsgId
+            $vmSize = [Microsoft.Azure.PowerShell.Cmdlets.ComputeFleet.Models.VMSizeProfile]::new()
+            $vmSize.Name = "Standard_D2s_v3"
+
+            $fleet = New-AzComputeFleet -Name $deleteViaIdentityFleetName `
+                -ResourceGroupName $resourceGroupName `
+                -SubscriptionId $env.SubscriptionId `
+                -Location $env.Location `
+                -ComputeProfileBaseVirtualMachineProfile $vmProfile `
+                -ComputeProfileComputeApiVersion "2024-03-01" `
+                -VMSizesProfile @($vmSize) `
+                -RegularPriorityProfileCapacity 1 `
+                -RegularPriorityProfileMinCapacity 1 `
+                -RegularPriorityProfileAllocationStrategy "LowestPrice" `
+                -Mode "Launch" `
+                -VMNamePrefix $vmNamePrefix `
+                -Tag @{ environment = "test" }
+
+            $fleet.Name | Should -Be $deleteViaIdentityFleetName
+            $fleet.ProvisioningState | Should -Be "Succeeded"
+
             Remove-AzComputeFleet -InputObject $fleet
-            $fleetList = Get-AzComputeFleet -SubscriptionId $env.SubscriptionId
-            $fleetList.Name | Should -Not -Contain $env.FleetName3
+
+            { Get-AzComputeFleet -Name $deleteViaIdentityFleetName `
+                -ResourceGroupName $resourceGroupName `
+                -SubscriptionId $env.SubscriptionId } | Should -Throw
         } | Should -Not -Throw
+    }
+
+    AfterAll {
+        if ($TestMode -ne 'playback') {
+            Remove-AzResourceGroup -Name $resourceGroupName -ErrorAction SilentlyContinue -Confirm:$false
+        }
     }
 }