+++ src/NetAppFiles/NetAppFiles.Test/ScenarioTests/VolumeTests.cs
@@ -60,8 +60,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Test.ScenarioTests.ScenarioTest
             TestRunner.RunTestScript("Test-VolumePipelines");
         }
 
-        //[Fact]
-        [Fact(Skip = "Requires AFEC not currently active")]
+        [Fact(Skip = "Doesn't work at the moment due to service side issue")]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void TestVolumeResetCifsOnNfsVolume()
         {