+++ src/NetAppFiles/NetAppFiles/Volume/NewNetAppFilesVolume.cs
@@ -303,6 +303,12 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Volume
 
         [Parameter(
             Mandatory = false,
+            HelpMessage = "Specifies whether the volume operates in Breakthrough Mode. Possible values include: 'Enabled', 'Disabled'")]
+        [PSArgumentCompleter("Enabled", "Disabled")]
+        public string BreakthroughMode { get; set; }
+
+        [Parameter(
+            Mandatory = false,
             HelpMessage = "A list of Availability Zones")]
         public string[] Zone { get; set; }
 
@@ -441,6 +447,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Volume
                     VolumeSpecName = VolumeSpecName,
                     PlacementRules = PlacementRule?.ToPlacementKeyValuePairs(),
                     EnableSubvolumes = EnableSubvolume.IsPresent ? EnableSubvolumes.Enabled : EnableSubvolumes.Disabled,
+                    BreakthroughMode = BreakthroughMode,
                     EncryptionKeySource = EncryptionKeySource,
                     KeyVaultPrivateEndpointResourceId = KeyVaultPrivateEndpointResourceId,
                     DeleteBaseSnapshot = DeleteBaseSnapshot,
@@ -524,6 +531,13 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Volume
                 SmbNonBrowsable = SmbNonBrowsable,
                 CoolAccessRetrievalPolicy = CoolAccessRetrievalPolicy
             };
+
+            var breakthroughModeProperty = volumeBody.GetType().GetProperty("BreakthroughMode");
+            if (breakthroughModeProperty != null)
+            {
+                breakthroughModeProperty.SetValue(volumeBody, BreakthroughMode);
+            }
+
             if (IsLargeVolume.IsPresent)
             {
                 volumeBody.IsLargeVolume = IsLargeVolume;