+++ src/Network/Network.Test/ScenarioTests/CortexTests.cs
@@ -46,6 +46,14 @@ namespace Commands.Network.Test.ScenarioTests
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         [Trait(Category.Owner, NrpTeamAlias.virtualwan)]
+        public void TestConnectionPolicyCRUD()
+        {
+            TestRunner.RunTestScript("Test-ConnectionPolicyCRUD");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.virtualwan)]
         public void TestStaticRoutesConfig()
         {
             TestRunner.RunTestScript("Test-StaticRoutesConfigCRUD");