+++ src/DnsResolver/DnsResolver/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 2.0.0
 * Added `Invoke-AzDnsResolverBulkDnsResolverDomainList`
 * [Upgraded code generator](https://go.microsoft.com/fwlink/?linkid=2340249)
 * Removed `ActionBlockResponseCode` from DNS security rule cmdlets