+++ src/EventHub/EventHub/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 5.6.0
+* Added parameter `IPAddressType` to cmdlets 'New-AzEventHubNamespace' and 'Set-AzEventHubNamespace'
+## Version 5.5.0
 * Added ChangeSafety Support
 
 ## Version 5.4.0