+++ src/NetAppFiles/NetAppFiles/ChangeLog.md
@@ -18,6 +18,14 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 1.4.0
+ * Added FileAccessLogs to `Get-AzNetAppFilesCache` output.
+ * Added Breakthrough Mode support for Azure NetApp Files volumes:
+     - Added `-BreakthroughMode` to `New-AzNetAppFilesVolume` and `New-AzNetAppFilesVolumeGroup`.
+     - Added `BreakthroughMode` to volume output in `Get-AzNetAppFilesVolume` and `Get-AzNetAppFilesVolumeGroup`
+
+## Version 1.3.0
 * Added ChangeSafety Support
 * Added FileAccessLogs to PSNetAppFilesCache ouput
 * Added Breaking Change notifications to Subvolume cmdlets in Az.NetAppFiles