+++ src/StorageSync/StorageSync/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 2.6.0
 * Added ChangeSafety Support
 
 ## Version 2.5.2